Extract syscall_gettimeofday() handler
